Add config for styles
